correcion

+++ Trabajo Practico Uno/funciones.c
@@ -57,28 +57,26 @@ float multiplicacion(float a,float b)
     resultado=a*b;
     printf("la multiplicacion de %.2f y %.2f es: %.2f\n",a,b,resultado);
 
-}
+
+    }
+
 
 
 int factorial(int a)
 {
-    int factor=(int)a;
+    int factor=a;
     int factorial=1;
     if(a<0)
     {
-
     printf("no se puede sacar factor numero negativo\n");
-
     }
 
     else
     {
-    for(int i=1; i<=(int)a; i++)
+    for(int i=1; i<=a; i++)
         {
             factorial *= i;
         }
          printf("El factorial de %d es: %d \n",a,factorial);
     }
 }
-
-
